# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /luno/transactions:
    get:
      operationId: listLunoTransactions
      tags: [luno]
      summary: Recent Luno wallet transactions
      parameters:
        - in: query
          name: asset
          required: false
          description: Filter by asset symbol (e.g. XBT, ZAR)
          schema: { type: string }
        - in: query
          name: accountId
          required: false
          description: Scope to a single linked Luno account (exchange-link id)
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: Transactions list (newest first)
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/LunoTransaction"
                required: [transactions]
...
  /luno/ticker:
    get:
      operationId: getLunoTicker
      tags: [luno]
      summary: Live ticker for a Luno pair (e.g. XBTZAR)
      parameters:
        - in: query
          name: pair
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Ticker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LunoTicker"
  /luno/tickers:
    get:
      operationId: getLunoTickers
      tags: [luno]
      summary: Batch live tickers for multiple Luno pairs
      description: >
        Quote N pairs in one round-trip. Per-pair failures (unsupported pair,
        upstream hiccup) are silently dropped from the result; clients must
        key by `pair` and tolerate missing entries.
      parameters:
        - in: query
          name: pairs
          required: true
          description: Comma-separated list of pairs, e.g. `XBTZAR,ETHZAR,SOLZAR`. Max 20.
          schema: { type: string }
      responses:
        "200":
          description: Tickers
          content:
            application/json:
              schema:
                type: object
                required: [tickers]
                properties:
                  tickers:
                    type: array
                    items:
                      $ref: "#/components/schemas/LunoTicker"
components:
  schemas:
...
    LunoTransaction:
      type: object
      properties:
        accountId: { type: string }
        accountName: { type: string }
        walletId: { type: string }
        asset: { type: string }
        rowIndex: { type: integer }
        ts: { type: string, format: date-time }
        amount:
          type: number
          description: Signed delta (positive = inflow, negative = outflow)
        balance:
          type: number
          description: Wallet balance after this transaction
        description: { type: string }
      required:
        - accountId
        - accountName
        - walletId
        - asset
        - rowIndex
        - ts
        - amount
        - balance
        - description
...
    LunoTicker:
      type: object
      properties:
        pair: { type: string }
        ask: { type: number }
        bid: { type: number }
        lastTrade: { type: number }
        rolling24hVolume: { type: number }
        asOf: { type: string, format: date-time }
      required: [pair, ask, bid, lastTrade, rolling24hVolume, asOf]
...